Fixes #137
* Enable Postgres using Spring Data primitives.
* Enable Liquibase

+++ ilpv4-connector-accounts/src/main/java/org/interledger/connector/accounts/AccountProviderSettings.java
@@ -76,7 +76,7 @@ public interface AccountProviderSettings {
    *
    * @return The maximum packet amount allowed by this account.
    */
-  Optional<BigInteger> getMaximumPacketAmount();
+  Optional<Long> getMaximumPacketAmount();
 
   /**
    * Defines whether the connector should maintain and enforce a balance for this account.
