+++ ilpv4-connector-server/src/main/java/com/sappenin/interledger/ilpv4/connector/server/spring/settings/properties/AccountProviderSettingsFromPropertyFile.java
@@ -6,7 +6,6 @@ import org.interledger.connector.accounts.AccountProviderSettings;
 import org.interledger.connector.accounts.AccountRelationship;
 import org.interledger.connector.link.LinkType;
 
-import java.math.BigInteger;
 import java.util.Map;
 import java.util.Optional;
 
@@ -34,7 +33,7 @@ public class AccountProviderSettingsFromPropertyFile implements AccountProviderS
 
   private LinkType LinkType;
   private Map<String, Object> customSettings = Maps.newConcurrentMap();
-  private Optional<BigInteger> maximumPacketAmount = Optional.empty();
+  private Optional<Long> maximumPacketAmount = Optional.empty();
 
   @Override
   public AccountProviderId getId() {
@@ -145,11 +144,11 @@ public class AccountProviderSettingsFromPropertyFile implements AccountProviderS
   }
 
   @Override
-  public Optional<BigInteger> getMaximumPacketAmount() {
+  public Optional<Long> getMaximumPacketAmount() {
     return maximumPacketAmount;
   }
 
-  public void setMaximumPacketAmount(Optional<BigInteger> maximumPacketAmount) {
+  public void setMaximumPacketAmount(Optional<Long> maximumPacketAmount) {
     this.maximumPacketAmount = maximumPacketAmount;
   }
 }
